requirements.txt
aiogram
pandas
apscheduler
selenium
mysql-connector-python
Pillow